+++ ving/docs/change-log.md
@@ -12,6 +12,7 @@ outline: deep
 * NOTE: useCurrentUserStore() has been renamed to useCurrentUser(), update your codebase appropriately.
 * NOTE: useNotifyStore() has been renamed to useNotify(), update your codebase appropriately.
 * NOTE: useSystemWideAlertStore() has been renamed to useSystemWideAlert(), update your codebase appropriately.
+* NOTE: useThrobberStore() has been renamed to useThrobber(), update your codebase appropriately.
 
 ## 2024-04-29
 * Implemented: reformat page generator to use panel components on view and edit #139